fix broken .fifth restoration
OY.

when loading a pre-1.3 collection that had broken fifths saved to it, rate and fifth-awareness were unsynced. so, when you played a pad that had an old "0.75 rate but .fifth is false" in it, recalling would break because it was trying to set the rate param with a value that didn't exist.